fix(work): restore TOC popover top padding, rename toc component
Blog fixed a padding gap after the TOCPopover bar by giving its prose
container pt-4; work's copy was missing that so the popover bar sat
flush against the first line of content. Consolidate work's px-4/pb-4
into p-4 with the top padding restored.

Also rename blog-toc.tsx to toc.tsx since TOCMinimap/TOCPopover are
shared between blog and work, not blog-specific.
