+++ src/app/(home)/work/[slug]/page.tsx
@@ -2,7 +2,6 @@ import { TOCProvider } from 'fumadocs-ui/components/toc'
 import type { Metadata } from 'next'
 import { notFound } from 'next/navigation'
 import { type ReactElement, ViewTransition } from 'react'
-import { TOCMinimap, TOCPopover } from '@/components/blog-toc'
 import { BlurImage } from '@/components/blur-image'
 import { Icons } from '@/components/icons/icons'
 import { WorkJsonLd } from '@/components/json-ld'
@@ -11,6 +10,7 @@ import { GitHubCode } from '@/components/mdx/github-code'
 import { PreviewButton } from '@/components/mdx/preview-button'
 import { Section } from '@/components/section'
 import { SectionBody } from '@/components/section-body'
+import { TOCMinimap, TOCPopover } from '@/components/toc'
 import { ViewAnimation } from '@/components/view-animation'
 import { description as homeDescription } from '@/constants/site'
 import { createMetadata, getWorkPageImage } from '@/lib/metadata'
@@ -127,7 +127,7 @@ export default async function Page(props: {
             initial={{ opacity: 0, translateY: -6 }}
             whileInView={{ opacity: 1, translateY: 0 }}
           >
-            <div className='prose prose-zinc dark:prose-invert prose-content min-w-0 flex-1 px-4 pb-4'>
+            <div className='prose prose-zinc dark:prose-invert prose-content min-w-0 flex-1 p-4'>
               <Mdx components={{ ...mdxComponents, GitHubCode }} />
             </div>
           </ViewAnimation>
